+++ packages/eyes-api/src/input/BatchInfo.ts
@@ -117,7 +117,11 @@ export class BatchInfoData implements Required<BatchInfo> {
     this.properties = properties
     return this
   }
-  addProperty(property: PropertyData): this {
+  addProperty(name: string, value: string): this
+  addProperty(prop: PropertyData): this
+  addProperty(propOrName: PropertyData | string, value?: string): this {
+    const property = utils.types.isString(propOrName) ? { name: propOrName, value } : propOrName
+    if (!this.properties) this.properties = []
     this.properties.push(property)
     return this
   }